+++ apps/kyb-app/src/hooks/useEndUserQuery/useEndUserQuery.ts
@@ -8,8 +8,11 @@ import {
 import { useQuery } from '@tanstack/react-query';
 import { HTTPError } from 'ky';
 import { useEffect } from 'react';
+import { useWorkflowId } from '@/common/hooks/useWorkflowId';
 
 export const useEndUserQuery = () => {
+  const workflowId = useWorkflowId();
+
   const {
     data: endUser,
     isLoading,
@@ -17,6 +20,8 @@ export const useEndUserQuery = () => {
     refetch,
   } = useQuery({
     ...collectionFlowQuerykeys.getEndUser(),
+    // @ts-ignore
+    enabled: !!workflowId,
     // @ts-ignore
     staleTime: Infinity as const,
   });