+++ packages/common/src/utils/collection-flow/build-collection-flow-url.ts
@@ -1,0 +1,15 @@
+export const buildCollectionFlowUrl = (
+  collectionFlowBaseUrl: string | undefined,
+  {
+    token,
+    workflowId,
+  }: {
+    token?: string | undefined;
+    workflowId: string | undefined;
+  },
+) =>
+  collectionFlowBaseUrl
+    ? `${collectionFlowBaseUrl}/collection-flow?workflowId=${workflowId ?? ''}${
+        token ? `&token=${token}` : ''
+      }`
+    : '';