+++ services/workflows-service/src/common/guards/workflow-guard/workflow-auth.decorator.ts
@@ -1,0 +1,6 @@
+import { applyDecorators, UseGuards } from '@nestjs/common';
+import { WorkflowAuthGuard } from './workflow-auth.guard';
+import { disableDefaultAuth } from '@/common/disable-default-auth';
+
+export const UseWorkflowAuthGuard = () =>
+  applyDecorators(UseGuards(WorkflowAuthGuard), disableDefaultAuth());