+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # kyb-app
 
+## 0.3.167
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/workflow-browser-sdk@0.6.115
+  - @ballerine/common@0.9.92
+
 ## 0.3.166
 
 ### Patch Changes